requirements.txt
pandas
google-cloud-language
